rewrite, tests, readme
Decided to rewrite the entire thing to utilize classes, for easier time with tests. Python appears to be very heavy on the usage of classes, and is tough to get away without using them, so I might as well embrace them more than I have been.

Included a README so GitHub can show a nice explanation on how to use the files, prettified using Markdown (because I am familiar with it)